requirements.txt
discord.py
transformers
torch
flask